Move interopKeypairs from lodestar-validator to lodestar-utils
