+++ packages/lodestar-validator/src/index.ts
@@ -5,5 +5,4 @@
 export * from "./validator";
 export * from "./options";
 export * from "./api";
-export * from "./interop";
 export * from "./slashingProtection";
