+++ packages/lodestar/test/unit/chain/genesis/genesis.test.ts
@@ -4,9 +4,8 @@ import {SecretKey, PublicKey} from "@chainsafe/bls";
 import {config} from "@chainsafe/lodestar-config/lib/presets/minimal";
 import {computeDomain, computeSigningRoot, DomainType} from "@chainsafe/lodestar-beacon-state-transition";
 import {DepositData, ValidatorIndex, DepositEvent, Eth1Block} from "@chainsafe/lodestar-types";
-import {ErrorAborted, WinstonLogger} from "@chainsafe/lodestar-utils";
+import {ErrorAborted, WinstonLogger, interopSecretKey} from "@chainsafe/lodestar-utils";
 import {toHexString} from "@chainsafe/ssz";
-import {interopSecretKey} from "@chainsafe/lodestar-validator";
 import {AbortController} from "abort-controller";
 import {IEth1Provider} from "../../../../src/eth1";
 import {GenesisBuilder} from "../../../../src/chain/genesis/genesis";
